+++ Project.toml
@@ -6,17 +6,13 @@ version = "0.3.0"
 ArnoldiMethod = "ec485272-7323-5ecc-a04f-4719b315124d"
 CommonSolve = "38540f10-b2f7-11e9-35d8-d573e4eb0ff2"
 ConstructionBase = "187b0558-2788-49d3-abe0-74a17ed4e7c9"
-DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 GeometryOps = "3251bfac-6a57-4b6d-aa61-ac1fef2975ab"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
-LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-ProgressLogging = "33c8b6b6-d38a-422a-b730-caa89a2f386c"
 Rasters = "a3a2b9e3-a471-40c9-b274-f788e487c689"
 SimpleWeightedGraphs = "47aef6b3-ad0c-573a-a1e2-d07658019622"
 SortTileRecursiveTree = "746ee33f-1797-42c2-866d-db2fce69d14d"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 WoodburyMatrices = "efce3f68-66dc-5838-9240-27a6d6f5f9b6"
 
 [weakdeps]
@@ -26,25 +22,33 @@ LinearSolve = "7ed4a6bd-45f5-4d41-b270-4a48e9bafcae"
 ConScapeLinearSolveExt = "LinearSolve"
 
 [compat]
+Aqua = "0.8"
+ArchGDAL = "0.10"
 ArnoldiMethod = "0.0.4, 0.4"
 CommonSolve = "0.2"
 ConstructionBase = "1.5.8"
-DelimitedFiles = "1"
 GeometryOps = "0.1.15"
 Graphs = "1"
-LaTeXStrings = "1.1"
+JET = "0.9"
+LinearAlgebra = "1"
 LinearSolve = "2.38.0"
 Plots = "1.4"
-ProgressLogging = "0.1"
 Rasters = "0.14"
+SafeTestsets = "0.1"
 SimpleWeightedGraphs = "1.1"
 SortTileRecursiveTree = "0.1.2"
+SparseArrays = "1"
+Test = "1"
 WoodburyMatrices = "1.0.0"
 julia = "1.10"
 
 [extras]
+Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
 ArchGDAL = "c9ce4bd3-c3d5-55b8-8973-c0e20141b8c3"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
+SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["ArchGDAL", "Plots"]
+test = ["Aqua", "ArchGDAL", "JET", "Plots", "SafeTestsets", "Test"]
